fix: Moved non rhino-specific classes into lower dependencies
