+++ DUI3-DX/Converters/Rhino/Speckle.Converters.Rhino7/ToHost/TopLevel/SpeckleFallbackToHostConversion.cs
@@ -1,4 +1,5 @@
-﻿using Speckle.Converters.Common;
+﻿using Objects.Other;
+using Speckle.Converters.Common;
 using Speckle.Converters.Common.Objects;
 using Speckle.Core.Models;
