+++ DUI3-DX/Sdk/Speckle.Converters.Common/Speckle.Converters.Common.csproj
@@ -7,6 +7,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\Objects\Objects\Objects.csproj" />
+    <ProjectReference Include="..\Speckle.Autofac\Speckle.Autofac.csproj" />
   </ItemGroup>
 
 </Project>
